requirements.txt
Flask==2.1.0
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
google-cloud-language==2.10.1
gunicorn==20.1.0